Update dependency com.maxmind.geoip2:geoip2 to v3

+++ backend/src/main/kotlin/org/tormap/service/RelayDetailsUpdateService.kt
@@ -62,7 +62,7 @@ class RelayDetailsUpdateService(
         val autonomousSystem = ipLookupService.lookupAutonomousSystem(this.address)
         if (autonomousSystem != null) {
             this.autonomousSystemName = autonomousSystem.autonomousSystemOrganization
-            this.autonomousSystemNumber = autonomousSystem.autonomousSystemNumber
+            this.autonomousSystemNumber = autonomousSystem.autonomousSystemNumber.toInt()
             relayDetailsRepositoryImpl.save(this)
             return true
         }
